Improve file organisation and reduce duplicate code

#9
