+++ app/src/main/java/com/leondeklerk/starling/library/folder/FolderFragment.kt
@@ -14,13 +14,12 @@ import com.leondeklerk.starling.data.MediaItem
 import com.leondeklerk.starling.data.MediaItemTypes
 import com.leondeklerk.starling.data.VideoItem
 import com.leondeklerk.starling.databinding.FragmentFolderBinding
-import com.leondeklerk.starling.gallery.GalleryAdapter
-import com.leondeklerk.starling.gallery.ui.GalleryFragmentDirections
+import com.leondeklerk.starling.media.MediaGalleryAdapter
 
 /**
  * [FolderFragment] is a gallery containing all items in a device folder.
  * It contains a recyclerView that contains all the images and videos.
- * Uses a [GalleryAdapter] assisted by [FolderViewModel] to display
+ * Uses a [MediaGalleryAdapter] assisted by [FolderViewModel] to display
  * [com.leondeklerk.starling.data.MediaItem]s.
  */
 class FolderFragment : Fragment() {
@@ -72,16 +71,16 @@ class FolderFragment : Fragment() {
         // Create item click listener
         val mediaItemClickListener = { item: MediaItem ->
             if (item.type == MediaItemTypes.VIDEO) {
-                val directions = GalleryFragmentDirections.actionNavigationGalleryToVideoActivity(item as VideoItem)
+                val directions = FolderFragmentDirections.actionFolderFragmentToVideoActivity(item as VideoItem)
                 findNavController().navigate(directions)
             } else if (item.type == MediaItemTypes.IMAGE) {
-                val directions = GalleryFragmentDirections.actionNavigationGalleryToImageActivity(item as ImageItem)
+                val directions = FolderFragmentDirections.actionFolderFragmentToImageActivity(item as ImageItem)
                 findNavController().navigate(directions)
             }
         }
 
         // Create a GalleryAdapter and add the data to it
-        val adapter = GalleryAdapter(mediaItemClickListener)
+        val adapter = MediaGalleryAdapter(mediaItemClickListener)
 
         folderViewModel.data.observe(
             viewLifecycleOwner,
