+++ app/src/main/res/navigation/navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/navigation_gallery"
-        android:name="com.leondeklerk.starling.gallery.ui.GalleryFragment"
+        android:name="com.leondeklerk.starling.gallery.GalleryFragment"
         android:label="@string/title_gallery"
         tools:layout="@layout/fragment_gallery" >
         <action
